Former-commit-id: 17ea98fbbf8a54782de00d3fa75eb58e93074664 [formerly 7e0e1399cc96b69cd55b2a4f4e45abecb46a7f98] [formerly 367e5de5813ce76d48c2642a94fbc78467b43a6d [formerly 21aa12bc5c6ccb84696a414222001b8cd2d1b749]]
Former-commit-id: dc9cbd33336d1892f63356b2754ffaa158ac67da [formerly 5827402930f56f217981fffd04f7a7466125a02f]
Former-commit-id: c83ef8efa9abcf0062a5166e0337a77d2a031971

+++ Frame/SDK/UtilityPlugin/AFCTimerModule.cpp
@@ -56,7 +56,7 @@ bool AFCTimerModule::AddSingleTimer(const std::string& name, const AFGUID& entit
     return mxTimerManager->AddSingleTimer(name, entity_id, interval_time, count, cb);
 }
 
-bool AFCTimerModule::AddForeverTimer(const std::string& name, const AFGUID& entity_id, const int64_t interval_time, TIMER_FUNCTOR_PTR cb)
+bool AFCTimerModule::AddForeverTimer(const std::string& name, const AFGUID& entity_id, const uint32_t interval_time, TIMER_FUNCTOR_PTR cb)
 {
     return mxTimerManager->AddForverTimer(name, entity_id, interval_time, cb);
 }
